+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/SequencesHolder.cs
@@ -136,7 +136,7 @@ internal class SequencesHolder
 
     public long KeyValueStoreEntryHistorySequence { get; set; }
 
-    public long ValidatorEmissionStatisticsSequence { get; set; }
+    public long ValidatorCumulativeEmissionHistorySequence { get; set; }
 
     public long NonFungibleSchemaHistorySequence { get; set; }