+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Migrations/20240610054430_InitialCreate.Designer.cs
@@ -80,7 +80,7 @@ using RadixDlt.NetworkGateway.PostgresIntegration.Models;
 namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
 {
     [DbContext(typeof(MigrationsDbContext))]
-    [Migration("20240528113230_InitialCreate")]
+    [Migration("20240610054430_InitialCreate")]
     partial class InitialCreate
     {
         /// <inheritdoc />
@@ -1714,7 +1714,7 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                     b.ToTable("validator_active_set_history");
                 });
 
-            modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.ValidatorEmissionStatistics", b =>
+            modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.ValidatorCumulativeEmissionHistory", b =>
                 {
                     b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
@@ -1730,6 +1730,10 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                     b.Property<long>("FromStateVersion")
                         .HasColumnType("bigint")
                         .HasColumnName("from_state_version");
+
+                    b.Property<long>("ParticipationInActiveSet")
+                        .HasColumnType("bigint")
+                        .HasColumnName("participation_in_active_set");
 
                     b.Property<long>("ProposalsMade")
                         .HasColumnType("bigint")
@@ -1747,9 +1751,7 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
 
                     b.HasIndex("ValidatorEntityId", "EpochNumber");
 
-                    NpgsqlIndexBuilderExtensions.IncludeProperties(b.HasIndex("ValidatorEntityId", "EpochNumber"), new[] { "ProposalsMade", "ProposalsMissed" });
-
-                    b.ToTable("validator_emission_statistics");
+                    b.ToTable("validator_cumulative_emission_history");
                 });
 
             modelBuilder.Entity("RadixDlt.NetworkGateway.PostgresIntegration.Models.ValidatorPublicKeyHistory", b =>